# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.userBody:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      role_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.AuditLog:
    properties:
      action:
        type: string
      actor_name:
        type: string
      actor_user_id:
        type: integer
      actor_username:
        type: string
      created_at:
        type: string
      http_method:
        type: string
      id:
        type: integer
      new_data:
        type: string
      old_data:
        type: string
      request_path:
        type: string
      resource_id:
        type: string
      resource_type:
        type: string
    type: object
  models.Employment:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Tokens:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    name: THS Armada Development
    url: https://github.com/armada-ths
  description: REST API powering the THS Armada career fair website (armada.nu) and
    its admin panel.
  title: ArmadaCMS API
  version: "1.0"
paths:
  /auditlogs:
    get:
      parameters:
      - description: Pagination range, e.g. [0,24]
        in: query
        name: range
        type: string
      - description: Sort, e.g. [\
        in: query
        name: sort
        type: string
      - description: 'Filter. Supported keys: action, resource_type, resource_id,
          actor_username, http_method, q (full-text)'
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Content-Range:
              description: auditlogs 0-24/1000
              type: string
          schema:
            items:
              $ref: '#/definitions/models.AuditLog'
            type: array
      security:
      - BearerAuth: []
      summary: List audit logs
      tags:
      - audit-logs
  /auditlogs/{id}:
    get:
      parameters:
      - description: AuditLog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuditLog'
        "404":
          description: Audit log entry not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get audit log by ID
      tags:
      - audit-logs
  /customusers:
    get:
      parameters:
      - description: Pagination range, e.g. [0,24]
        in: query
        name: range
        type: string
      - description: Sort, e.g. [\
        in: query
        name: sort
        type: string
      - description: Filter, e.g. {\
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Content-Range:
              description: users 0-24/100
              type: string
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.userBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid body
          schema:
            type: string
        "500":
          description: Create failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /customusers/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: User not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.userBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid data
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
...
  /employments:
    get:
      parameters:
      - description: Pagination range, e.g. [0,24]
        in: query
        name: range
        type: string
      - description: Sort, e.g. [\
        in: query
        name: sort
        type: string
      - description: Filter, e.g. {\
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Content-Range:
              description: employments 0-24/10
              type: string
          schema:
            items:
              $ref: '#/definitions/models.Employment'
            type: array
      summary: List employment types
      tags:
      - employments
    post:
      consumes:
      - application/json
      parameters:
      - description: Employment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Employment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Employment'
        "400":
          description: Invalid body
          schema:
            type: string
        "500":
          description: Create failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create employment
      tags:
      - employments
  /employments/{id}:
    delete:
      parameters:
      - description: Employment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            type: string
        "404":
          description: employment not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete employment
      tags:
      - employments
    get:
      parameters:
      - description: Employment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employment'
        "404":
          description: employment not found
          schema:
            type: string
      summary: Get employment by ID
      tags:
      - employments
    put:
      consumes:
      - application/json
      parameters:
      - description: Employment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated employment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Employment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employment'
        "400":
          description: Invalid data
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update employment
      tags:
      - employments
  /eventroevents:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
        "500":
          description: Failed to fetch from Eventro
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Sync & list events from Eventro
      tags:
      - eventro
  /eventroexhibitors:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Exhibitor'
            type: array
        "500":
          description: Failed to fetch from Eventro
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Sync & list exhibitors from Eventro
      tags:
      - eventro
  /eventromembers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Profile'
            type: array
        "500":
          description: Failed to fetch from Eventro
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Sync & list member profiles from Eventro
      tags:
      - eventro
  /eventrorecruitments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Sync complete
          schema:
            type: string
        "500":
          description: Failed to fetch from Eventro
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Sync recruitment data from Eventro
      tags:
      - eventro
  /events:
    get:
      parameters:
      - description: Pagination range, e.g. [0,24]
        in: query
        name: range
        type: string
      - description: Sort, e.g. [\
        in: query
        name: sort
        type: string
      - description: Filter, e.g. {\
        in: query
        name: filter
        type: string
      - description: If true, only return events with show=true
        in: query
        name: public
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Content-Range:
              description: events 0-24/100
              type: string
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
      summary: List events
      tags:
      - events
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Event name
        in: formData
        name: name
        required: true
        type: string
      - description: Location
        in: formData
        name: location
        required: true
        type: string
      - description: Start time (RFC3339)
        in: formData
        name: eventStart
        required: true
        type: string
      - description: End time (RFC3339)
        in: formData
        name: eventEnd
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Food information
        in: formData
        name: food
        type: string
      - description: Registration required
        in: formData
        name: registrationRequired
        type: boolean
      - description: Fee information
        in: formData
        name: fee
        type: string
      - description: Signup link URL
        in: formData
        name: signupLink
        type: string
      - description: Max capacity
        in: formData
        name: eventMaxCapacity
        type: integer
      - description: Registration end time (RFC3339)
        in: formData
        name: registrationEnd
        type: string
      - description: Whether to show the event publicly
        in: formData
        name: show
        type: boolean
      - description: Event image
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Create failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - events
  /events/{id}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - events
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "404":
          description: Event not found
          schema:
            type: string
      summary: Get event by ID
      tags:
      - events
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event name
        in: formData
        name: name
        type: string
      - description: Location
        in: formData
        name: location
        type: string
      - description: Start time (RFC3339)
        in: formData
        name: eventStart
        type: string
      - description: End time (RFC3339)
        in: formData
        name: eventEnd
        type: string
      - description: Whether to show the event publicly
        in: formData
        name: show
        type: boolean
      - description: Event image
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update event
      tags:
      - events
  /exhibitors:
    get:
      parameters:
      - description: Pagination range, e.g. [0,24]
        in: query
        name: range
        type: string
      - description: Sort, e.g. [\
        in: query
        name: sort
        type: string
      - description: Filter, e.g. {\
        in: query
        name: filter
        type: string
      - description: Return all exhibitors without pagination, sorted by tier
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Content-Range:
              description: exhibitors 0-24/100
              type: string
          schema:
            items:
              $ref: '#/definitions/models.Exhibitor'
            type: array
      summary: List exhibitors
      tags:
      - exhibitors
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Company name
        in: formData
        name: name
        required: true
        type: string
      - description: Company type
        in: formData
        name: type
        required: true
        type: string
      - description: Fair location
        in: formData
        name: fairLocation
        required: true
        type: string
      - description: Tier (Standard/Bronze/Silver/Gold)
        in: formData
        name: tier
        type: string
      - description: About the company
        in: formData
        name: about
        type: string
      - description: Purpose/mission
        in: formData
        name: purpose
        type: string
      - description: Company website URL
        in: formData
        name: companyWebsite
        type: string
      - description: Climate compensation
        in: formData
        name: climateCompensation
        type: boolean
      - description: City/cities
        in: formData
        name: cities
        type: string
      - description: Logo image (JPG/PNG/WEBP/GIF)
        in: formData
        name: file
        type: file
      - description: JSON array of program IDs, e.g. [1,2,3]
        in: formData
        name: programs
        type: string
      - description: JSON array of industry IDs
        in: formData
        name: industries
        type: string
      - description: JSON array of employment IDs
        in: formData
        name: employments
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Exhibitor'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Create failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create exhibitor
      tags:
      - exhibitors
...
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "401":
          description: Wrong username or password
          schema:
            type: string
        "500":
          description: Login failed
          schema:
            type: string
      summary: Login
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer JWT token obtained from POST /api/v1/login. Format: "Bearer
      <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"